requirements.txt
# Python version requirement
# This project requires Python 3.10 or higher

# Core Composio dependencies
composio>=0.10.6,<2.0.0
composio-langgraph>=0.10.6,<2.0.0

# LangChain and LangGraph dependencies
langchain-core>=1.2.7,<3.0.0
langchain-openai>=1.1.7,<3.0.0
langgraph>=1.0.6,<2.0.0

# Environment management
python-dotenv>=1.2.1,<2.0.0

# Note: Standard library packages (os, datetime, random) are built-in and don't need installation
#
# To generate a full locked requirements file with all transitive dependencies, run:
# pip freeze > requirements-lock.txt

annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
composio==0.10.6
composio-client==1.25.0
composio-langgraph==0.10.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-openai==1.1.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
openai==2.15.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
Pysher==1.0.8
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
websocket-client==1.9.0
xxhash==3.6.0
zstandard==0.25.0